+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 
+## [0.7.3] - 2026-05-01
+
+### Changed
+
+- (BREAKING) GitGraph is split into TrackMap, TrackLayout, and LabelMap.
+  WARNING: This was a major refactoring so the code may still have some bugs.
+
+
 ## [0.7.2] - 2026-04-14
 
 ### Added